requirements.txt
certifi==2022.12.7
charset-normalizer==2.0.12
filelock==3.10.0
ftfy==6.1.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pip==23.0.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
python-dateutil==2.8.2
pytz==2022.7.1
qtmodern==0.2.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.0
setuptools==67.6.0
six==1.16.0
sympy==1.11.1
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
wheel==0.40.0
